type cast to np.int32

+++ layered_vision/filters/fill.py
@@ -55,7 +55,7 @@ class FillFilter(AbstractOptionalChannelFilter):
         if config.poly_points is not None:
             image_size = get_image_size(image_array)
             width_height_tuple = (image_size.width, image_size.height,)
-            poly_points = (config.poly_points * width_height_tuple).astype(np.int)
+            poly_points = (config.poly_points * width_height_tuple).astype(np.int32)
             return cv2.fillPoly(
                 np.array(image_array, dtype=np.float_),
                 pts=[poly_points],
